+++ Banks/Accounts/DepositAccount.cs
@@ -5,35 +5,35 @@ namespace Banks.AccountTypes
 {
     public class DepositAccount : AccountBuilder
     {
-        public override void SetPercent(float percent)
+        public override void SetPercent(double percent)
         {
         }
 
-        public override void SetDepositPercent(float percent)
+        public override void SetDepositPercent(double percent)
         {
             Account.Percent = percent;
         }
 
-        public override void SetStartBalance(float startBalance)
+        public override void SetStartBalance(double startBalance)
         {
             Account.IncreaseMoney(startBalance);
         }
 
-        public override void SetMaxWithdraw(float amount)
+        public override void SetMaxWithdraw(double amount)
         {
             Account.MaxWithdraw = amount;
         }
 
-        public override void SetMaxRemittance(float amount)
+        public override void SetMaxRemittance(double amount)
         {
             Account.MaxRemittance = amount;
         }
 
-        public override void SetCreditLimit(float creditLimit)
+        public override void SetCreditLimit(double creditLimit)
         {
         }
 
-        public override void SetCommission(float commission)
+        public override void SetCommission(double commission)
         {
         }
